+++ src/Helpers/VsHelpers.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.VisualStudio.Shell;
+
+namespace TypeScriptCompileOnSave
+{
+    public static class VsHelpers
+    {
+        public static TReturnType GetService<TServiceType, TReturnType>()
+        {
+            return (TReturnType)ServiceProvider.GlobalProvider.GetService(typeof(TServiceType));
+        }
+    }
+}