+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Road map
 
-- [ ] nothing yet...
+- [x] Write to the status bar when transpiling
+- [x] Only run transpiler when needed
 
 Features that have a checkmark are complete and available for
 download in the